+++ src/TrashLib.Tests/Sonarr/ReleaseProfileUpdaterTest.cs
@@ -1,6 +1,7 @@
 ﻿using NSubstitute;
 using NUnit.Framework;
 using Serilog;
+using TrashLib.Sonarr;
 using TrashLib.Sonarr.Api;
 using TrashLib.Sonarr.Config;
 using TrashLib.Sonarr.ReleaseProfile;
@@ -16,6 +17,7 @@ namespace TrashLib.Tests.Sonarr
             public IReleaseProfileGuideParser Parser { get; } = Substitute.For<IReleaseProfileGuideParser>();
             public ISonarrApi Api { get; } = Substitute.For<ISonarrApi>();
             public ILogger Logger { get; } = Substitute.For<ILogger>();
+            public ISonarrCompatibility Compatibility { get; } = Substitute.For<ISonarrCompatibility>();
         }
 
         [Test]
@@ -23,7 +25,7 @@ namespace TrashLib.Tests.Sonarr
         {
             var context = new Context();
 
-            var logic = new ReleaseProfileUpdater(context.Logger, context.Parser, context.Api);
+            var logic = new ReleaseProfileUpdater(context.Logger, context.Parser, context.Api, context.Compatibility);
             logic.Process(false, new SonarrConfiguration());
 
             context.Parser.DidNotReceive().GetMarkdownData(Arg.Any<ReleaseProfileType>());
@@ -40,7 +42,7 @@ namespace TrashLib.Tests.Sonarr
                 ReleaseProfiles = new[] {new ReleaseProfileConfig {Type = ReleaseProfileType.Anime}}
             };
 
-            var logic = new ReleaseProfileUpdater(context.Logger, context.Parser, context.Api);
+            var logic = new ReleaseProfileUpdater(context.Logger, context.Parser, context.Api, context.Compatibility);
             logic.Process(false, config);
 
             context.Parser.Received().ParseMarkdown(config.ReleaseProfiles[0], "theMarkdown");