+++ src/TrashLib/Sonarr/ISonarrCompatibility.cs
@@ -1,0 +1,10 @@
+namespace TrashLib.Sonarr
+{
+    public interface ISonarrCompatibility
+    {
+        bool SupportsNamedReleaseProfiles { get; }
+        bool ArraysNeededForReleaseProfileRequiredAndIgnored { get; }
+        string InformationalVersion { get; }
+        string MinimumVersion { get; }
+    }
+}